add comments and fill readme

+++ src/main/java/mazurek/it/climbingcompetition/users/UserRepository.java
@@ -5,13 +5,14 @@ import org.springframework.data.repository.query.Param;
 import org.springframework.data.rest.core.annotation.RepositoryRestResource;
 
 import java.util.List;
-
+// This will be AUTO IMPLEMENTED by Spring into a Bean called userRepository
+// CRUD refers Create, Read, Update, Delete
 @RepositoryRestResource(collectionResourceRel = "users", path = "users")
-public interface UserRepository extends CrudRepository<User, Long> {
+public interface UserRepository extends CrudRepository<User, Integer> {
 
 //    SPRING sam dostarcza URLe do searcha po zdefiniowanych funkcjach
 //    np. http://localhost:8080/users/search/findUsersByPhone?phone=668164072
-//    nie jest to estetyczne ale działa
+
     List<User> findUsersByLastName(@Param("name") String name);
     List<User> findUsersByPhone(@Param("phone") String phone);
